requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=9.5.0

# YOLOv5
ultralytics>=8.0.0

# ONNX
onnx>=1.14.0
onnxsim>=0.4.33
onnxruntime-gpu>=1.16.0

# Data / Math
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Demo
gradio>=4.0.0

# Optional: TensorRT (Jetson pre-installed, or install manually)
# tensorrt>=8.6.0
# pycuda>=2022.2.2

# Evaluation
pycocotools>=2.0.6

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
requests>=2.31.0
